Extract solver body management into new `SolverBodyPlugin`
